FileAddressBookDao.getPersonCountBySex implemented by using Java 8 Stream API, tests added. PersonMapper implemented and tested.

+++ src/main/java/com/gumtree/user/dao/AddressBookDao.java
@@ -5,7 +5,7 @@ import com.gumtree.user.model.Sex;
 
 public interface AddressBookDao {
 
-	int getPersonCountBySex(Sex sex);
+	long getPersonCountBySex(Sex sex);
 
 	Person getOldestPerson();
