+++ src/main/java/com/gumtree/user/service/AddressBookService.java
@@ -6,7 +6,7 @@ import com.gumtree.user.model.Sex;
 
 public interface AddressBookService {
 
-	int getPersonCountBySex(Sex sex);
+	long getPersonCountBySex(Sex sex);
 
 	Person getOldestPerson();
